Wrote some functional tests
This includes the initial setup for using the test-environment.

As you will see we have a fairly large travis.yml-file. This is because we
have a lot of dependencies in grunt to generate our assest. As the test should
be ran on an environment as close to the production as possible I tried to use
the full grunt build, but on travis we can't install some packages, therefor
we only install only what is required.

*Warning*: when updating the versions in packages.json make sure you update
them in .travis.yml-file as well.