Avoid database access during assetic dump.
When doing an assetic dump, assetic will go trough all the templates and
compile all the assets.

Our framework contains some twig extensions that have our database as a
dependency (for example our settings mechanism). When accessing our
templates, this means our database needs to have all it's needed
configuration.

When Doctrine (DBAL) is instantiated, some variables need to be known.
One of them is the version of MySQL you are running (because this is
needed to make some functionality work). If it's not provided in the
config, doctrine will do a call to the database to fetch this version.

This commit makes sure this is not needed anymore. This means we can do
assetic:dump without the need of having a working database connection.

See https://github.com/kriswallsmith/assetic/issues/681 for more info.