Fixed issue with gulp.dest
For some reason dependencies for task will not be complete before starting the actual task.
The gulp-documentation sucks, as they state: "An array of tasks to be executed and completed before your task will run.", but this is not correct.
But they added a note, which in their opinion resolves all issues...